[JENKINS-27374] Make existing tests pass. (Applying script-security, migrating to jenkins-1.509, maven-hpi-plugin-1.117)
