+++ src/main/resources/index.jelly
@@ -1,3 +1,4 @@
+<?jelly escape-by-default='true'?>
 <div>
   Extensible Choice Parameter provides several ways to retrieve choices,
   including the way to share choices among all jobs.
