+++ src/main/resources/jp/ikedam/jenkins/plugins/extensible_choice_parameter/TextareaChoiceListProvider/config.jelly
@@ -21,6 +21,7 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 -->
+<?jelly escape-by-default='true'?>
 <j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout" xmlns:t="/lib/hudson" xmlns:f="/lib/form" xmlns:m="/jp/ikedam/jenkins/plugins/extensible_choice_parameter/taglib">
     <f:entry title="${%Choices}" field="choiceListText" help="/help/parameter/choice-choices.html">
         <f:textarea />
