+++ src/test/java/jp/ikedam/jenkins/plugins/extensible_choice_parameter/SystemGroovyChoiceListProviderJenkinsTest.java
@@ -33,6 +33,7 @@ import hudson.util.ListBoxModel;
 import java.util.Arrays;
 import java.util.List;
 
+import org.jenkinsci.plugins.scriptsecurity.scripts.ScriptApproval;
 import org.junit.Rule;
 import org.junit.Test;
 import org.jvnet.hudson.test.JenkinsRule.WebClient;
@@ -84,7 +85,7 @@ public class SystemGroovyChoiceListProviderJenkinsTest
         
         // Proper script
         {
-            ListBoxModel ret = descriptor.doFillDefaultChoiceItems(null, properScript, false);
+            ListBoxModel ret = descriptor.doFillDefaultChoiceItems(null, properScript, true, false);
             assertEquals("Script returned an unexpected list", properScriptReturn.size() + 1, ret.size());
             for(int i = 0; i < properScriptReturn.size(); ++i)
             {
@@ -94,7 +95,7 @@ public class SystemGroovyChoiceListProviderJenkinsTest
         
         // Non-string list script
         {
-            ListBoxModel ret = descriptor.doFillDefaultChoiceItems(null, nonstringScript, false);
+            ListBoxModel ret = descriptor.doFillDefaultChoiceItems(null, nonstringScript, true, false);
             assertEquals("Script returned an unexpected list", nonstringScriptReturn.size() + 1, ret.size());
             for(int i = 0; i < nonstringScriptReturn.size(); ++i)
             {
@@ -104,49 +105,49 @@ public class SystemGroovyChoiceListProviderJenkinsTest
         
         // non-list script
         {
-            ListBoxModel ret = descriptor.doFillDefaultChoiceItems(null, nonlistScript, false);
+            ListBoxModel ret = descriptor.doFillDefaultChoiceItems(null, nonlistScript, true, false);
             assertEquals("Script returning non-list must return an empty list", 1, ret.size());
         }
         
         // Empty list script
         {
-            ListBoxModel ret = descriptor.doFillDefaultChoiceItems(null, emptyListScript, false);
+            ListBoxModel ret = descriptor.doFillDefaultChoiceItems(null, emptyListScript, true, false);
             assertEquals("Script must return an empty list", 1, ret.size());
         }
         
         // Null script
         {
-            ListBoxModel ret = descriptor.doFillDefaultChoiceItems(null, nullScript, false);
+            ListBoxModel ret = descriptor.doFillDefaultChoiceItems(null, nullScript, true, false);
             assertEquals("Script with null must return an empty list", 1, ret.size());
         }
         
         // emptyScript
         {
-            ListBoxModel ret = descriptor.doFillDefaultChoiceItems(null, emptyScript, false);
+            ListBoxModel ret = descriptor.doFillDefaultChoiceItems(null, emptyScript, true, false);
             assertEquals("empty script must return an empty list", 1, ret.size());
         }
         
         // blankScript
         {
-            ListBoxModel ret = descriptor.doFillDefaultChoiceItems(null, blankScript, false);
+            ListBoxModel ret = descriptor.doFillDefaultChoiceItems(null, blankScript, true, false);
             assertEquals("blank script must return an empty list", 1, ret.size());
         }
         
         // Syntax broken script
         {
-            ListBoxModel ret = descriptor.doFillDefaultChoiceItems(null, syntaxBrokenScript, false);
+            ListBoxModel ret = descriptor.doFillDefaultChoiceItems(null, syntaxBrokenScript, true, false);
             assertEquals("Syntax-broken-script must return an empty list", 1, ret.size());
         }
         
         // exceptionScript
         {
-            ListBoxModel ret = descriptor.doFillDefaultChoiceItems(null, exceptionScript, false);
+            ListBoxModel ret = descriptor.doFillDefaultChoiceItems(null, exceptionScript, true, false);
             assertEquals("Script throwing an exception must return an empty list", 1, ret.size());
         }
         
         // null
         {
-            ListBoxModel ret = descriptor.doFillDefaultChoiceItems(null, null, false);
+            ListBoxModel ret = descriptor.doFillDefaultChoiceItems(null, null, true, false);
             assertEquals("null must return an empty list", 1, ret.size());
         }
     }
@@ -158,31 +159,31 @@ public class SystemGroovyChoiceListProviderJenkinsTest
         
         // Proper script
         {
-            FormValidation formValidation = descriptor.doTest(null, properScript, false);
+            FormValidation formValidation = descriptor.doTest(null, properScript, true, false);
             assertEquals("Test for proper script must succeed", FormValidation.Kind.OK, formValidation.kind);
         }
         
         // Syntax broken script
         {
-            FormValidation formValidation = descriptor.doTest(null, syntaxBrokenScript, false);
+            FormValidation formValidation = descriptor.doTest(null, syntaxBrokenScript, true, false);
             assertEquals("Test for broken script must fail", FormValidation.Kind.ERROR, formValidation.kind);
         }
         
         // Script raising an exception
         {
-            FormValidation formValidation = descriptor.doTest(null, exceptionScript, false);
+            FormValidation formValidation = descriptor.doTest(null, exceptionScript, true, false);
             assertEquals("Test for script raising an exception must fail", FormValidation.Kind.ERROR, formValidation.kind);
         }
         
         // Script returning non-list
         {
-            FormValidation formValidation = descriptor.doTest(null, nonlistScript, false);
+            FormValidation formValidation = descriptor.doTest(null, nonlistScript, true, false);
             assertEquals("Test for script returning non-list must fail", FormValidation.Kind.ERROR, formValidation.kind);
         }
         
         // Script returning null
         {
-            FormValidation formValidation = descriptor.doTest(null, nullScript, false);
+            FormValidation formValidation = descriptor.doTest(null, nullScript, true, false);
             assertEquals("Test for script retuning null must fail", FormValidation.Kind.ERROR, formValidation.kind);
         }
     }
@@ -302,6 +303,9 @@ public class SystemGroovyChoiceListProviderJenkinsTest
     @Test
     public void testVariables() throws Exception
     {
+        ScriptApproval.get().approveSignature("method hudson.model.PersistenceRoot getRootDir");
+        ScriptApproval.get().approveSignature("method java.io.File getAbsolutePath");
+        ScriptApproval.get().approveSignature("method hudson.model.Item getFullName");
         FreeStyleProject p = j.createFreeStyleProject();
         p.addProperty(new ParametersDefinitionProperty(new ExtensibleChoiceParameterDefinition(
                 "test",
@@ -325,6 +329,7 @@ public class SystemGroovyChoiceListProviderJenkinsTest
     @Test
     public void testProjectVariable() throws Exception
     {
+        ScriptApproval.get().approveSignature("method hudson.model.Item getFullName");
         FreeStyleProject p = j.createFreeStyleProject();
         CaptureEnvironmentBuilder ceb = new CaptureEnvironmentBuilder();
         p.addProperty(new ParametersDefinitionProperty(new ExtensibleChoiceParameterDefinition(
@@ -368,5 +373,78 @@ public class SystemGroovyChoiceListProviderJenkinsTest
         assertEquals(0, cli.execute("build", p.getFullName(), "-s"));
         j.assertBuildStatusSuccess(p.getLastBuild());
         assertEquals(p.getFullName(), ceb.getEnvVars().get("test"));
+    }
+    
+    @Test
+    public void testSystemGroovyChoiceListProvider_scriptText()
+    {
+        // simple value
+        {
+            String scriptText = "abc";
+            SystemGroovyChoiceListProvider target = new SystemGroovyChoiceListProvider(scriptText, null);
+            assertEquals("SystemGroovyChoiceListProvider must preserve scriptText", scriptText, target.getScriptText());
+        }
+        
+        // value with spaces
+        {
+            String scriptText = "  abc  \n  cdf  ";
+            SystemGroovyChoiceListProvider target = new SystemGroovyChoiceListProvider(scriptText, null);
+            assertEquals("SystemGroovyChoiceListProvider must preserve scriptText", scriptText, target.getScriptText());
+        }
+        
+        // empty
+        {
+            String scriptText = "";
+            SystemGroovyChoiceListProvider target = new SystemGroovyChoiceListProvider(scriptText, null);
+            assertEquals("SystemGroovyChoiceListProvider must preserve scriptText", scriptText, target.getScriptText());
+        }
+        
+        // blank
+        {
+            String scriptText = "   ";
+            SystemGroovyChoiceListProvider target = new SystemGroovyChoiceListProvider(scriptText, null);
+            assertEquals("SystemGroovyChoiceListProvider must preserve scriptText", scriptText, target.getScriptText());
+        }
+        
+        // null
+        {
+            String scriptText = null;
+            SystemGroovyChoiceListProvider target = new SystemGroovyChoiceListProvider(scriptText, null);
+            assertEquals("SystemGroovyChoiceListProvider must preserve scriptText", scriptText, target.getScriptText());
+        }
+    }
+    
+    @Test
+    public void testSystemGroovyChoiceListProvider_defaultChoice()
+    {
+        String scriptText = "abc";
+        
+        // a value
+        {
+            String defaultChoice = "some value";
+            SystemGroovyChoiceListProvider target = new SystemGroovyChoiceListProvider(scriptText, defaultChoice);
+            assertEquals("a value", defaultChoice, target.getDefaultChoice());
+        }
+        
+        // null
+        {
+            String defaultChoice = null;
+            SystemGroovyChoiceListProvider target = new SystemGroovyChoiceListProvider(scriptText, defaultChoice);
+            assertEquals("null", defaultChoice, target.getDefaultChoice());
+        }
+        
+        // empty
+        {
+            String defaultChoice = "";
+            SystemGroovyChoiceListProvider target = new SystemGroovyChoiceListProvider(scriptText, defaultChoice);
+            assertEquals("empty", defaultChoice, target.getDefaultChoice());
+        }
+        
+        // blank
+        {
+            String defaultChoice = "  ";
+            SystemGroovyChoiceListProvider target = new SystemGroovyChoiceListProvider(scriptText, defaultChoice);
+            assertEquals("blank", defaultChoice, target.getDefaultChoice());
+        }
     }
 }
